django-testing/django_testing/requirements.txt
django~=3.2.7
djangorestframework~=3.12.4
psycopg2-binary

django-filter~=2.4.0
pytest~=6.2.5